832-refactor: Mentorship folder (#845)
* refactor: 832 - mentorship folder

* refactor: 832 - rename folders and files in mentorship

* refactor: 832 - rename mentoship type

* refactor: 832 - additional refactoring

---------
