+++ src/views/mentorship/mentorship.tsx
@@ -1,12 +1,12 @@
-import { MentorshipHero } from './mentorship-hero/ui/mentorship-hero';
-import { DetailsMentorship } from './ui/details/details';
-import { MentorActivities } from './ui/mentor-activities';
-import { MentorshipCourses } from './ui/mentorship-courses/mentorship-courses';
-import { MentorsAfterRegister } from '@/views/mentorship/ui/mentors-after-register';
 import { Benefits } from '@/widgets/benefits';
 import { MentorsDocs } from '@/widgets/mentors-docs';
 import { MentorsFeedback } from '@/widgets/mentors-feedback';
 import { MentorsRegister } from '@/widgets/mentors-register';
+import { MentorActivities } from '@/widgets/mentorship-activities';
+import { MentorshipAfterRegister } from '@/widgets/mentorship-after-register';
+import { MentorshipCourses } from '@/widgets/mentorship-courses/mentorship-courses';
+import { MentorshipDetails } from '@/widgets/mentorship-details/mentorship-details';
+import { MentorshipHero } from '@/widgets/mentorship-hero';
 import { MentorshipCourse, benefitMentorshipMentors, mentorsFeedbackData } from 'data';
 
 type MentorshipProps = {
@@ -18,7 +18,7 @@ export const Mentorship = ({ mentorshipData, courses = false }: MentorshipProps)
   return (
     <>
       <MentorshipHero lang={mentorshipData.lang} />
-      <DetailsMentorship description={mentorshipData.details} lang={mentorshipData.lang} />
+      <MentorshipDetails description={mentorshipData.details} lang={mentorshipData.lang} />
       {!mentorshipData.title && <Benefits {...benefitMentorshipMentors} />}
       <MentorActivities
         activitiesTitle={mentorshipData.activitiesTitle}
@@ -27,7 +27,7 @@ export const Mentorship = ({ mentorshipData, courses = false }: MentorshipProps)
       {courses && <MentorshipCourses />}
       <MentorsFeedback mentorsFeedback={mentorsFeedbackData} />
       <MentorsRegister lang={mentorshipData.lang} />
-      <MentorsAfterRegister lang={mentorshipData.lang} />
+      <MentorshipAfterRegister lang={mentorshipData.lang} />
       <MentorsDocs
         mentorDocsLink={mentorshipData.links.mentorDocs}
         courseDocsLink={mentorshipData.links.courseDocs}
