--- src/views/mentorship/ui/mentors-after-register/index.ts
@@ -1,1 +1,0 @@
-export { MentorsAfterRegister } from './ui/mentors-after-register';
