+++ src/widgets/mentorship-after-register/ui/mentorship-after-register.module.scss
